Struct named_pipe::PipeOptions

pub struct PipeOptions {
    // some fields omitted
}

Options and flags which can be used to configure how a pipe is created.

This builder exposes the ability to configure how a ConnectingServer is created.

Builder defaults:

open_mode - Duplex

in_buffer - 65536

out_buffer - 65536

first - true

Methods

impl PipeOptions

fn new<T: AsRef<OsStr>>(name: T) -> PipeOptions

fn first(&mut self, val: bool) -> &mut PipeOptions

Is this instance (or instances) will be first for this pipe name? Defaults to true.

fn open_mode(&mut self, val: OpenMode) -> &mut PipeOptions

Open mode for pipe instance. Defaults to Duplex.

fn in_buffer(&mut self, val: u32) -> &mut PipeOptions

Input buffer size for pipe instance. Defaults to 65536

fn out_buffer(&mut self, val: u32) -> &mut PipeOptions

Output buffer size for pipe instance. Defaults to 65536.

fn single(&self) -> Result<ConnectingServer>

Creates single instance of pipe with this options.

fn multiple(&self, num: u32) -> Result<Vec<ConnectingServer>>

Creates multiple instances of pipe with this options.