Struct named_pipe::PipeServer

pub struct PipeServer {
    // some fields omitted
}

Pipe server connected to a client.

Methods

impl PipeServer

fn unwrap(self) -> Result<ConnectingServer>

This function will flush buffers and disconnect server from client. Then will start waiting for a new client.

fn read_async<'a, 'b>(&'a mut self, buf: &'b mut [u8]) -> Result<ReadHandle<'a, Self>>

Initializes asyncronous read opeation.

fn read_async_owned(self, buf: Vec<u8>) -> Result<ReadHandle<'static, Self>>

Initializes asyncronous read operation and takes ownership of buffer and server.

fn write_async<'a, 'b>(&'a mut self, buf: &'b [u8]) -> Result<WriteHandle<'a, Self>>

Initializes asyncronous write operation.

fn write_async_owned(self, buf: Vec<u8>) -> Result<WriteHandle<'static, Self>>

Initializes asyncronous write operation and takes ownership of buffer and server.