Crate named_pipe

Named-Pipe is a wrapper for overlapped (asyncronous) IO of Windows's named pipes.

Description

You can use wait or wait_all to select(2)-like wait for multiple pending IO operations (which is read/write from/to PipeServer/PipeClient or waiting for new client).

Or you can use ConnectingServer::wait or io::Read and io::Write implementaions for PipeServer and PipeClient for syncronous communication.

For better understanding please refer to [Named Pipes documentation on MSDN] (https://www.google.com/search?q=msdn+named+pipes&ie=utf-8&oe=utf-8).

Usage

To create new pipe instance use PipeOptions structure.

To connect to a pipe server use PipeClient structure.

Structs

ConnectingServer
Pipe instance waiting for new client. Can be used with wait and [wait_all] (fn.wait_all.html) functions.

PipeClient
Pipe client connected to a server.

PipeIoHandles

PipeIoObj

PipeOptions
Options and flags which can be used to configure how a pipe is created.

PipeServer
Pipe server connected to a client.

ReadHandle
Pending read operation. Can be used with wait and [wait_all] (fn.wait_all.html) functions.

WriteHandle
Pending write operation. Can be used with wait and [wait_all] (fn.wait_all.html) functions.

Enums

OpenMode

Traits

PipeIo
This trait used for genericity.

Functions

wait
This function will wait for first finished io operation and return it's index in list.